+++ CHANGELOG.rst
@@ -54,6 +54,9 @@ Breaking API Changes
 * finagle-netty4: The type of context of a `ChannelTransport` has been changed from a
   `Netty4Context` to a `ChannelTransportContext`. ``PHAB_ID=D205794``
 
+* finagle-netty4: `c.t.f.netty4.param.useUnpoolledByteBufAllocator` flag has been removed. There is
+  no good reason to opt-out of a more efficient, pooled allocator. ``PHAB_ID=D212097``
+
 * finagle-thrift: `DeserializeCtx` became `ClientDeserializeCtx` for client side response
   classification, add `ServerDeserializeCtx` to handle server side response classification.
   ``PHAB_ID=D196032``